refactor(dungeon-builder): move ThumbHarness out of src/concepts/
ThumbHarness.tsx (the dev-only ?thumbGlb= R3F thumbnail-baking harness,
gated in App.tsx the same way PlaytestHarness's ?encounterId= is) has
no dependency on anything dungeon-builder-specific -- it just happens
to be the tool that baked that concept's palette thumbnails. Having
App.tsx import it from src/concepts/dungeon-builder/thumbs/ made a
concept folder a dependency of the app shell (graduation audit item).
Moved to src/dev/ThumbHarness.tsx, alongside DevPerfProbe.tsx (the
existing src/dev/ convention). The thumbs/*.png assets themselves stay
in src/concepts/dungeon-builder/thumbs/ -- paletteData.ts's
import.meta.glob('./thumbs/*.png', ...) is relative to paletteData.ts's
own location, which didn't move, so this was unaffected.

Updated App.tsx's import and paletteData.ts's doc-comment pointer to
the new path.

typecheck/lint/tests clean (1772/1772); production build also verified
clean -- the palette thumbnail glob still resolves all 13 baked PNGs
correctly post-move.
